refactor(MSW): replace MSWComponent with MSWProvider
